feat: Componentize dropdown from OnSaleItems
